Bound storage by `Value` trait
