+++ dunge/src/lib.rs
@@ -12,6 +12,7 @@ mod state;
 pub mod storage;
 pub mod texture;
 pub mod uniform;
+pub mod value;
 pub mod vertex;
 
 #[cfg(feature = "winit")]
